Rip out local caching of successful passwords
Hopefully pause.perl.org is much more reliable than 7 years ago.  We can
re-add it if necessary, but this reduces complexity and means old
passwords won't continue to work between the time you change it on PAUSE
and the time you next login to rt.cpan.org with the new password.